+++ src/main/java/kitchenpos/common/PurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.infra;
+package kitchenpos.common;
 
 public interface PurgomalumClient {
     boolean containsProfanity(String text);
