+++ src/main/java/kitchenpos/infra/DefaultPurgomalumClient.java
@@ -1,5 +1,6 @@
 package kitchenpos.infra;
 
+import kitchenpos.common.PurgomalumClient;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
