+++ src/test/java/kitchenpos/infra/FakePurgomalumClient.java
@@ -2,6 +2,7 @@ package kitchenpos.infra;
 
 import java.util.Arrays;
 import java.util.List;
+import kitchenpos.common.PurgomalumClient;
 
 public class FakePurgomalumClient implements PurgomalumClient {
     private static final List<String> profanities;
