+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -2,7 +2,7 @@ package kitchenpos.menu.application;
 
 import kitchenpos.infra.FakePurgomalumClient;
 import kitchenpos.menu.domain.InMemoryMenuRepository;
-import kitchenpos.infra.PurgomalumClient;
+import kitchenpos.common.PurgomalumClient;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menu.domain.MenuRepository;
